Add Database configuration for SQL reports
For reports that are SQL based - users should be able to specify a
configuration for the database that points to a different database
that uses the same schema. This can be useful for testing the same
report works in environments like Development, Production.
Currently, the configuration applies to ALL report files but in future
we will support having different config per report - if necessary
and perhaps support NoSQL datasources aswell.